requirements/local.txt
asgiref==3.5.2
Django==4.0.6
django-ckeditor==6.4.2
django-js-asset==2.0.0
django-model-utils==4.2.0
Pillow==9.2.0
psycopg2==2.9.3
sqlparse==0.4.2
tzdata==2022.1
Unipath==1.1